requirements.txt
google-generativeai
python-dotenv
streamlit
PIL
io
requests